Start tidying CSR unit
Constants that are defined in contranomy_types are now used instead
of literals in csrunit. Casts to std_logic_vector are removed, they
don't seem to be needed by GHDL but add a lot of line noise, e.g.

  std_logic_vector'(std_logic_vector'("00"))

instead of

  "00"

Next step is to see if there are any places where a std_logic_vector
would be better replaced by a more informative type.